requirements.txt
PyYAML
mock
nose
nose-cov
setuptools
